refactor handleSubmit into UserInput
