+++ src/app/chat/page.tsx
@@ -1,15 +1,12 @@
 'use client';
 
 import { useState, useEffect } from 'react';
-import { useRouter } from "next/navigation";
 import { UserInput } from "@/components/UserInput";
 import { supabase } from '@/utils/supabase'
 import { Session } from '@supabase/supabase-js'
 import { logger } from '@/utils/logger'
 
 export default function ChatPage() {
-    const router = useRouter();
-    const [isLoading, setIsLoading] = useState(false);
     const [jobDescription, setJobDescription] = useState("");
     const [session, setSession] = useState<Session | null>(null)
 
@@ -26,30 +23,6 @@ export default function ChatPage() {
         return <div>loading..</div>//router.push('/login')
     }
 
-    const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-            e.preventDefault();
-            setIsLoading(true);
-    
-            if (e.currentTarget.elements && "jobDescription" in e.currentTarget.elements) {
-                const jobDescriptionElement = e.currentTarget.elements.jobDescription as HTMLTextAreaElement;
-                const jobDescriptionValue = jobDescriptionElement.value;
-    
-                // Send job description to the server
-                const response = await fetch("/api/jobs", {
-                    method: "POST",
-                    headers: {"Content-Type": "application/json"},
-                    body: JSON.stringify({description: jobDescriptionValue}),
-                });
-    
-                const {jobId} = await response.json();
-                router.push(`/job/${jobId}`); // Redirect to the job page
-            } else {
-                console.error("jobDescription element not found in form.");
-                setIsLoading(false);
-                logger.debug(`Ignore this error, it's a quic fix for non-used isLoading state ${isLoading}`);
-            }
-        };
-
   return (
     <main>
     {/* Right Side - Input Form */}
@@ -58,7 +31,6 @@ export default function ChatPage() {
         <UserInput
             jobDescription={jobDescription}
             onJobDescriptionChange={setJobDescription}
-            handleSubmit={handleSubmit}
         />
     </div>
     </main>
