+++ src/app/page.tsx
@@ -1,7 +1,6 @@
 'use client'
 
 import React, { useState } from "react";
-import { useRouter } from "next/navigation";
 import { UserInput } from '@/components/UserInput';
 import { useLoading } from '@/context/LoadingContext';
 import Image from 'next/image';
@@ -35,33 +34,8 @@ const testimonials = [
 ];
 
 export default function Home() {
-    const router = useRouter();
     const [jobDescription, setJobDescription] = useState("");
     const { setIsLoading, isLoading } = useLoading();
-
-
-    const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-        e.preventDefault();
-        setIsLoading(true);
-
-        if (e.currentTarget.elements && "jobDescription" in e.currentTarget.elements) {
-            const jobDescriptionElement = e.currentTarget.elements.jobDescription as HTMLTextAreaElement;
-            const jobDescriptionValue = jobDescriptionElement.value;
-
-            // Send job description to the server
-            const response = await fetch("/api/jobs", {
-                method: "POST",
-                headers: {"Content-Type": "application/json"},
-                body: JSON.stringify({description: jobDescriptionValue}),
-            });
-
-            const {jobId} = await response.json();
-            router.push(`/job/${jobId}`); // Redirect to the job page
-        } else {
-            console.error("jobDescription element not found in form.");
-            setIsLoading(false);
-        }
-    };
 
     return (
         <main
@@ -112,7 +86,6 @@ export default function Home() {
                             <UserInput
                                 jobDescription={jobDescription}
                                 onJobDescriptionChange={setJobDescription}
-                                handleSubmit={handleSubmit}
                             />
                         </div>
                     </div>
